Clean up inline functions, add utility scripts, misc.
- Add `WI_INLINE` (expands to `static inline`)  macro with `always_inline` attribute (on GCC) to `wi_util.h`
- Replace all `static inline` with `WI_INLINE`
- Fix implicit integer conversions in `wi_disasm.c` functions (`_X_instr`)
- Move `wi_state_append_error` from inline to real function (it couldn't be inlined anyway because of varargs)
- Make `wi_state_check_arity` inline
- Optimize `wi_value_is_bool` and `wi_value_is_falsy`
- Add debug/release mode to `xmake.lua`
- Update benchmark results
- Add utility scripts: `build.sh`, `pack.sh`, and `profile.sh`

+++ src/core/wi_box.c
@@ -132,7 +132,7 @@ wi_prototype_instr_size(struct wi_prototype* prototype, int offset) {
     }
 
     /* WI_OP_PUSH_CLOSURE */
-    uint16_t             constant          = bytes[offset + 1] << 8 | bytes[offset + 2];
+    uint16_t             constant          = (uint16_t)(bytes[offset + 1] << 8 | bytes[offset + 2]);
     struct wi_prototype* closure_prototype = wi_value_as_prototype(prototype->constants.data[constant]);
     return 3 + closure_prototype->upvalue_count * 2;
 }
